fix: CI lint エラー + Docker /designs プリレンダリングエラー修正
- E741: 曖昧な変数名 `l` → `lang` にリネーム（dynamic_scholar_block.py + テスト）
- F401: 未使用 import `ALLOWED_LANGUAGES` を publisher_tools.py から削除
- /designs ページの useSearchParams() を Suspense バウンダリで包み、
  Docker ビルド時のプリレンダリングエラーを回避
